u64 改为 u128

+++ src/lib.rs
@@ -4,7 +4,7 @@
 use std::cmp::Ordering;
 
 /// 更相减损术
-pub fn subtraction(mut x: u64, mut y: u64) -> u64 {
+pub fn subtraction(mut x: u128, mut y: u128) -> u128 {
     loop {
         match x.cmp(&y) {
             Ordering::Less => y -= x,
@@ -15,7 +15,7 @@ pub fn subtraction(mut x: u64, mut y: u64) -> u64 {
 }
 
 /// 更相减损术 除2
-pub fn subtraction2(mut x: u64, mut y: u64) -> u64 {
+pub fn subtraction2(mut x: u128, mut y: u128) -> u128 {
     let mut i = 1;
     while x % 2 == 0 && y % 2 == 0 {
         x /= 2;
@@ -33,7 +33,7 @@ pub fn subtraction2(mut x: u64, mut y: u64) -> u64 {
 }
 
 /// 辗转相除法
-pub fn euclidean(mut x: u64, mut y: u64) -> u64 {
+pub fn euclidean(mut x: u128, mut y: u128) -> u128 {
     while x * y != 0 {
         match x.cmp(&y) {
             Ordering::Less => {
